+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 3.0.0
+* replace '__' prefix for custom extensions with 'x-' prefix
+* add some modelFuncs to facilitate the usage of openApi models in templates
+
 # 2.1.2
 * fix a bug in handling single type models w/o any title